+++ app/src/main/java/jp/techacademy/akihiro/ogawa/calcapp/SecondActivity.java
@@ -13,8 +13,8 @@ public class SecondActivity extends AppCompatActivity {
         setContentView(R.layout.activity_second);
 
         Intent intent = getIntent();
-        int value1 = intent.getIntExtra("VALUE1", 0);
-        int value2 = intent.getIntExtra("VALUE2", 0);
+        Double value1 = intent.getDoubleExtra("VALUE1", 0);
+        Double value2 = intent.getDoubleExtra("VALUE2", 0);
         int value3 = intent.getIntExtra("VALUE3", 0);
 
         if (value3 == 1 ){
